requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
colorama==0.4.6
pip==23.3.1
pycparser==2.21
pycryptodome==3.15.0
pywin32==305.1
setuptools==68.2.2
tqdm==4.65.0
typing_extensions==4.9.0
wheel==0.41.2
